Extract Sun subfolder to an embedded package
This commit also fixes a couple of breakages in the ExampleMunicipality. Every effort has been made to keep this backwards compatible, including making sure that the Sun prefab's id/hash is still the same and thus any application using that prefab will not break.
