+++ Packages/Netherlands3D/Sun/Runtime/Scripts/DateTimeStorage.cs
@@ -1,26 +1,23 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 using Netherlands3D.Events;
+using UnityEngine;
 
 public class DateTimeStorage : MonoBehaviour
 {
-
     [SerializeField] private DateTimeContainer dateTimeContainer;
     [SerializeField] private DateTimeEvent onDateTimeStored;
     [SerializeField] private DateTimeEvent onDateTimeLoaded;
     private DateTime receivedDateTime;
 
-
     private void Awake()
     {
-        if(dateTimeContainer == null)
+        if (dateTimeContainer == null)
         {
             // The DateTimeContainer can be predefined, but if it doesn't exist, one will be created on Awake.
             dateTimeContainer = ScriptableObject.CreateInstance<DateTimeContainer>();
         }
     }
+
     public void ReceiveDateTime(DateTime dateTime)
     {
         // This function can receive the dateTime every frame and it saves it, but doesn't keep it in the storage.
@@ -30,7 +27,7 @@ public class DateTimeStorage : MonoBehaviour
     public void StoreCurrentDateTime()
     {
         dateTimeContainer.SetValue(receivedDateTime);
-        if(onDateTimeStored != null)
+        if (onDateTimeStored != null)
         {
             onDateTimeStored.InvokeStarted(dateTimeContainer.Value);
         }
@@ -38,11 +35,9 @@ public class DateTimeStorage : MonoBehaviour
 
     public void LoadStoredDateTime()
     {
-        if(onDateTimeLoaded != null)
+        if (onDateTimeLoaded != null)
         {
             onDateTimeLoaded.InvokeStarted(dateTimeContainer.Value);
         }
     }
-
-
 }
