+++ Packages/Netherlands3D/Sun/Runtime/Scripts/DateTimeToString.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -12,6 +10,4 @@ public class DateTimeToString : MonoBehaviour
     {
         dateTimeText.text = dateTime.ToString();
     }
-
-
 }
