+++ Packages/Netherlands3D/Sun/Runtime/Scripts/VariableContainers/BaseVariableContainer.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
