+++ Packages/Netherlands3D/Sun/Runtime/Scripts/VariableContainers/DateTimeContainer.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "New DateTimeContainer", menuName = "DataContainers/DateTimeContainer")]
@@ -9,5 +7,5 @@ public class DateTimeContainer : BaseVariableContainer<DateTime>
     public override string ToString()
     {
         return Value.ToString();
-    } 
+    }
 }
