+++ Packages/eu.netherlands3d.sun/Runtime/Prefabs/Sun.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b1a31d57791238b4097b78a08ca6d032
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
