+++ Packages/eu.netherlands3d.sun/Runtime/Scripts/eu.netherlands3d.sun.Runtime.asmdef
@@ -1,12 +1,9 @@
 {
-    "name": "Netherlands3D.Sun.Runtime",
+    "name": "eu.netherlands3d.sun.Runtime",
     "rootNamespace": "Netherlands3D.Sun",
     "references": [
-        "GUID:9ca88277166fea644a31588b4f943acc",
-        "GUID:15fc0a57446b3144c949da3e2b9737a9",
         "GUID:19c45768ddcd67943afbf0cb4e4ff0b3",
-        "GUID:6fc4e9e40fe9cfd41bc3559197fc47b1",
-        "GUID:ab927a7f481373442a8ee26564bcd9ef"
+        "GUID:15fc0a57446b3144c949da3e2b9737a9"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
