+++ Packages/eu.netherlands3d.sun/Runtime/Scripts/eu.netherlands3d.sun.Runtime.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ab927a7f481373442a8ee26564bcd9ef
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
